pofcc-375 - Created common root module; moved decentralisation source into the common module; use common module in api/ and src/
